+++ spec/functional/mongoid/relations/referenced/many_to_many_spec.rb
@@ -1203,6 +1203,14 @@ describe Mongoid::Relations::Referenced::ManyToMany do
           it "adds the document to the target" do
             person.preferences.count.should == 1
           end
+
+          it "does not duplicate documents" do
+            person.reload.preferences.count.should eq(1)
+          end
+
+          it "does not duplicate ids" do
+            person.reload.preference_ids.count.should eq(1)
+          end
         end
       end
     end